+++ back-end/src/main/java/com/invest/dto/ContaNewDTO.java
@@ -1,5 +1,7 @@
 package com.invest.dto;
 
+import java.math.BigDecimal;
+
 import com.invest.entity.enums.TipoConta;
 
 import lombok.AllArgsConstructor;
@@ -16,7 +18,7 @@ public class ContaNewDTO {
 
 	private Integer numero;
 	private Integer agencia;
-	private double saldo;
+	private BigDecimal saldo;
 	private Integer banco;
 	private TipoConta tipo;
 	private String cpf;
